% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDuplicateID.R
\name{checkDuplicateID}
\alias{checkDuplicateID}
\title{Function to check duplicate ID variables}
\usage{
checkDuplicateID(data, id)
}
\arguments{
\item{data}{S dataset}

\item{id}{ID column corresponding to the unique subject ID.}
}
\description{
Function to check duplicate ID variables
}
