% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageEntropy.R
\name{codeEntropy}
\alias{codeEntropy}
\title{Helper function to convert code language entropy}
\usage{
codeEntropy(data, id_quo, cols_quo, contextName, base)
}
\arguments{
\item{data}{The dataset}

\item{id_quo}{Quoted ID variable corresponding to the unique subject ID}

\item{cols_quo}{Quoted columns of the dataset that be used to compute entropy}

\item{contextName}{Title for the resulting entropy column}

\item{base}{Base of the logarithm}
}
\description{
Helper function to convert code language entropy
}
