% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert2prop.R
\name{conv2prop}
\alias{conv2prop}
\title{Helper function to convert to Likert to proportions}
\usage{
conv2prop(data, id_quo, cols_quo, minLikert = 1)
}
\arguments{
\item{data}{A dataset that includes columns of likert data.}

\item{id_quo}{Quoted ID variable corresponding to the unique subject ID}

\item{cols_quo}{Quoted columns of the dataset that contain the likert data.}

\item{minLikert}{The minimum possible value of your likert index / scale. Typically 1.}
}
\description{
Helper function to convert to Likert to proportions
}
