% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageEntropy.R
\name{entropy}
\alias{entropy}
\title{Simple entropy function}
\usage{
entropy(x, base = 2)
}
\arguments{
\item{x}{A vector of discrete probabilities}

\item{base}{Base of the logarithm. Shannon Entropy uses base 2. Other common bases are e (natural) and 10 (hartley).}
}
\description{
Given a vector of discrete probabilities, this function will output an
Entropy score (a measure of uncertainty). This function doesn't do any type
of error checking (e.g., that proportions add up to 1). However, if all
proportions add up to 0, the function will output NA.
}
\examples{
prob_table <- data.frame(event=c("A","B"), prob=c(.25,.75))
entropy(prob_table$prob)
}
\keyword{entropy}
\keyword{shannon}
