% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageEntropy.R
\name{languageEntropy}
\alias{languageEntropy}
\title{Language entropy function}
\usage{
languageEntropy(data, id, ..., entropyName, respType = "proportion",
  minLikert = 1)
}
\arguments{
\item{data}{A dataset that includes columns which will be converted into entropy}

\item{id}{ID variable corresponding to the unique subject ID}

\item{...}{Columns of the dataset that will be converted to entropy.}

\item{respType}{The response type of cols. Should either by "likert", "percent", or "proportion". Default is "propotion".}

\item{minLikert}{If using liker scales, specify the minimum possible likert value, typically 1.}
}
\description{
This function applies the simple entropy function to specified columns within
a dataset for each individual subject. The user should supply the column that
identifies the unique subject ID (id) as well as the columns of the dataset
that contain the data to be converted into entropy (cols). The user must
specify the type of data contained in the specified columns (i.e., whether
the data are likert data, percentage data, or proportion data). Likert and
percentage data are converted to proportions. Note that the probabilities
should sum to 1.
}
