% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert2prop.R
\name{likert2prop}
\alias{likert2prop}
\title{Function to convert Likert scale data to proportions}
\usage{
likert2prop(data, id, ..., colsList = NULL, minLikert = 1)
}
\arguments{
\item{data}{Your dataset}

\item{id}{ID column corresponding to the unique subject ID.}

\item{...}{Columns of the dataset that contain the Likert data to be
converted to proportions. If using this argument, proportions for each column will be
computed by summing together all columns. If you want to specify groups of
columns to be converted independently, you may specify them in the colsList
argument instead}

\item{colsList}{A list of grouped columns. E.g.,
list(c("L1Home","L2Home","L3Home"), c("L1Work","L2Work","L3Work")). Totals
will be computed separately for each group.}

\item{minLikert}{The minimum possible value of your Likert index / scale.
Typically 1. This is used in rebaselining Likert values to 0.}
}
\description{
Function to convert Likert scale data to proportions
}
\examples{
library(languageEntropy)
data(entropyExData) # load example data

# convert Likert scale data to proportions

## first for the home context
entropyExData <- likert2prop(entropyExData, sub, L1Home, L2Home, L3Home)
print(entropyExData)

# next for the work context
entropyExData <- likert2prop(entropyExData, sub, L1Work, L2Work, L3Work)
print(entropyExData)

entropyExData <- likert2prop(entropyExData, sub, L1PercentUse, L2PercentUse, L3PercentUse)
print(entropyExData)

# alternatively, you can convert home and work at the same time
# by passing home and work as separate vectors within a list
data(entropyExData) # reload example data
entropyExData <- likert2prop(entropyExData, sub, colsList = list(c("L1Home", "L2Home", "L3Home"), c("L1Work", "L2Work", "L3Work")))
print(entropyExData)
}
