% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent2prop.R
\name{percent2prop}
\alias{percent2prop}
\title{Function to convert to percentages to proportions}
\usage{
percent2prop(data, ...)
}
\arguments{
\item{data}{A dataset that includes columns of percentage data.}

\item{...}{Columns of the dataset that contain the percentage data.}

\item{id}{ID column corresponding to the unique subject ID.}
}
\description{
Function to convert to percentages to proportions
}
\examples{
library(languageEntropy)
data(entropyExData) # load example data

# convert percentage data to proportions

entropyExData <- percent2prop(entropyExData, L1PercentUse, L2PercentUse, L3PercentUse)
print(entropyExData)
}
