% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageEntropy.R
\name{codeEntropy}
\alias{codeEntropy}
\title{Helper function to convert code language entropy}
\usage{
codeEntropy(data, id_quo, cols_quo, sphereName, base)
}
\arguments{
\item{data}{A dataset that includes columns of likert data.}

\item{id_quo}{Quoted ID variable corresponding to the unique subject ID}

\item{cols_quo}{Quoted columns of the dataset that contain the likert data.}

\item{sphereName}{Title for the resulting entropy column.}

\item{base}{Base of the logarithm.}
}
\description{
Helper function to convert code language entropy
}
