% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageEntropy.R
\name{languageEntropy}
\alias{languageEntropy}
\title{languageEntropy function}
\usage{
languageEntropy(data, id, ..., sphereName = NULL, colsList = NULL,
  base = 2)
}
\arguments{
\item{data}{A dataset that includes columns which will be converted into
entropy}

\item{id}{ID variable corresponding to the unique subject ID}

\item{...}{Columns of the dataset that will be converted to entropy.}

\item{sphereName}{A title you give for the resulting entropy column. Leave
this NULL if you use colsList.}

\item{colsList}{A list of grouped columns. Entropy will be computed
separately for each group. E.g., list(Home=c("L1Home","L2Home","L3Home"),
Work=c("L1Work","L2Work","L3Work")).}

\item{base}{Base of the logarithm. Default is Shannon Entropy with base 2.
Other common bases are e (natural) and 10 (hartley).}
}
\description{
This function applies the simple entropy function to specified columns within
a dataset for each individual subject. The user should supply the column that
identifies the unique subject ID (id) as well as the columns of the dataset
that contain the data to be converted into entropy (...).
}
\details{
Note that entropy will be computed from proportions and these proportions
should sum to 1 within a usage sphere. There are various helper functions to
help you convert different types of data to propotions (likert2prop() and
percent2prop()).
}
\examples{
library(languageEntropy)
data(entropyExDataProp) # load example data with proportions

# first for the home sphere
entropyExDataProp <- languageEntropy(entropyExDataProp, sub, L1Home_prop, L2Home_prop, L3Home_prop, sphereName = "Home")
print(entropyExDataProp)

# next for the work sphere
entropyExDataProp <- languageEntropy(entropyExDataProp, sub, L1Work_prop, L2Work_prop, L3Work_prop, sphereName = "Work")
print(entropyExDataProp)

# finally for percent/proportion usage sphere
entropyExDataProp <- languageEntropy(entropyExDataProp, sub, L1PercentUse_prop, L2PercentUse_prop, L3PercentUse_prop, sphereName = "PercentUse")
print(entropyExDataProp)

# alternatively, you can deal with home, work and percent at the same time
# by passing data for these spheres as separate vectors within a list
data(entropyExDataProp) #reload the data
entropyExData <- languageEntropy(entropyExData, sub,
colsList = list(Home=c("L1Home_prop", "L2Home_prop", "L3Home_prop"),
                Work=c("L1Work_prop","L2Work_prop","L3Work_prop"),
                PercentUse=c("L1PercentUse_prop", "L2PercentUse_prop", "L3PercentUse_prop")))
}
