% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert2prop.R
\name{likert2prop}
\alias{likert2prop}
\title{Function to convert Likert to proportions}
\usage{
likert2prop(data, id, ..., colsList = NULL, minLikert = 1)
}
\arguments{
\item{data}{A dataset that includes columns of likert data.}

\item{id}{ID column corresponding to the unique subject ID.}

\item{...}{Columns of the dataset that contain the likert data. If using this argument, proportions will be computed using all columns specified here as the total. If you want to specify several groups of columns, where each group is totaled separately, you may specify them in the colsList agrument below.}

\item{colsList}{A list of grouped columns. E.g., list(c("L1Home","L2Home","L3Home"), c("L1Work","L2Work","L3Work")). Totals will be computed separately for each group.}

\item{minLikert}{The minimum possible value of your likert index / scale. Typically 1.}
}
\description{
Function to convert Likert to proportions
}
\examples{
library(languageEntropy)
data(entropyExData) # load example data

# convert Likert scale data to proportions

## first for the home sphere
entropyExData <- likert2prop(entropyExData, sub, L1Home, L2Home, L3Home)
print(entropyExData)

# next for the work sphere
entropyExData <- likert2prop(entropyExData, sub, L1Work, L2Work, L3Work)
print(entropyExData)

entropyExData <- likert2prop(entropyExData, sub, L1PercentUse, L2PercentUse, L3PercentUse)
print(entropyExData)

# alternatively, you can deal with home and work at the same time
# by passing home and work as separate vectors within a list
data(entropyExData) # reload example data
entropyExData <- likert2prop(entropyExData, sub, colsList = list(c("L1Home", "L2Home", "L3Home"), c("L1Work", "L2Work", "L3Work")))
print(entropyExData)
}
